We implemented code from Bilbot for the Frobot OpMode.
