+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/lib/mechanisms/FakeIntakeImpl.java
@@ -1,0 +1,8 @@
+package org.firstinspires.ftc.teamcode.lib.mechanisms;
+
+public class FakeIntakeImpl implements Intake {
+    @Override
+    public void run() {
+        //nothing (:
+    }
+}
