+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/lib/mechanisms/FakeLauncherImpl.java
@@ -1,0 +1,8 @@
+package org.firstinspires.ftc.teamcode.lib.mechanisms;
+
+public class FakeLauncherImpl implements Launcher {
+    @Override
+    public void run() {
+        //nothing lol
+    }
+}
