+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/lib/mechanisms/FakeTransferSystemImpl.java
@@ -1,0 +1,9 @@
+package org.firstinspires.ftc.teamcode.lib.mechanisms;
+
+public class FakeTransferSystemImpl implements TransferSystem {
+
+    @Override
+    public void run() {
+        //nothing!!!
+    }
+}
